requirements.txt
Django==4.2.6
django-ipware==5.0.1
djangorestframework==3.14.0
dnspython==2.5.0
docopt==0.6.2
ecdsa==0.19.0
email-validator==2.1.0.post1
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
httplib2==0.22.0
PyMySQL==1.1.0
pyodbc==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ipware==3.0.0
python-jose==3.3.0
python-jwt==2.0.1
python-multipart==0.0.6
sqlparse==0.4.4

